began moving Teacher class to its own module to clean up the main file
